refactor: move email exception from common to email
