+++ src/email/exceptions/index.ts
